+++ libs/core/src/handlers/events.ts
@@ -15,6 +15,7 @@ export enum SeelenEvent {
   MediaDevices = "media::devices",
   MediaInputs = "media-inputs",
   MediaOutputs = "media-outputs",
+  MediaWaveform = "media::waveform",
   NetworkDefaultLocalIp = "network-default-local-ip",
   NetworkAdapters = "network-adapters",
   NetworkInternetConnection = "network-internet-connection",