+++ src/background/exposed.rs
@@ -246,6 +246,7 @@ pub fn register_invoke_handler(app_builder: Builder<Wry>) -> Builder<Wry> {
     use crate::modules::fonts::infrastructure::*;
     use crate::modules::media::devices::infrastructure::*;
     use crate::modules::media::players::infrastructure::*;
+    use crate::modules::media::waveform::infrastructure::*;
     use crate::modules::monitors::brightness::infrastructure::*;
     use crate::modules::monitors::infrastructure::*;
     use crate::modules::network::infrastructure::*;