+++ src/background/modules/media/mod.rs
@@ -1,2 +1,3 @@
 pub mod devices;
 pub mod players;
+pub mod waveform;